+++ common/include/pcl/experimental/executor/trait/can_require.hpp
@@ -16,20 +16,6 @@
 namespace pcl {
 namespace executor {
 
-namespace detail {
-
-/**
- * \brief Checks if the given Executor supports the Property
- *
- *\details This is checked through a template variable static_query_v which is
- * provided by all properties
- */
-template <typename Executor, typename Property>
-using contains_property = std::is_same<
-    std::remove_const_t<decltype(Property::template static_query<Executor>::value)>,
-    Property>;
-} // namespace detail
-
 /**
  * \brief Enforces a specified Property on an Executor. A new executor instance
  * which implements that property is created and returned.
